Use a single ubi.js file, and reference that it is source from UBI plugin repo
